move ensurePg & instertInitialData to Global
